Add total sold product api

+++ src/main/java/shop/haui_megatech/domain/dto/product/TopSoldProductResponseDTO.java
@@ -9,6 +9,6 @@ public record TopSoldProductResponseDTO(
         Float newPrice,
         Integer totalSold,
         String mainImageUrl,
-        Long revenue
+        Double revenue
 ) {
 }
